Move code used only by the extension into its scope and out of the common package
