+++ common/src/index.ts
@@ -1,5 +1,5 @@
 /*!
- * Copyright 2018-2019 VMware, Inc.
+ * Copyright 2018-2021 VMware, Inc.
  * SPDX-License-Identifier: MIT
  */
 
@@ -17,10 +17,8 @@ export { default as Logger, LogChannel } from "./logger"
 export { AutoWire } from "./di"
 export * from "./event"
 export * from "./maven"
-export * from "./ts"
 export * from "./platform"
 export * from "./types"
 export * from "./rest"
-export * from "./hierarchy"
 
 export { di, proc, promise, uri, validate }
