--- common/src/ts/index.ts
@@ -1,6 +1,0 @@
-/*!
- * Copyright 2018-2019 VMware, Inc.
- * SPDX-License-Identifier: MIT
- */
-
-export * from "./VrotscCliProxy"
