+++ extension/src/client/util/index.ts
@@ -1,0 +1,6 @@
+/*!
+ * Copyright 2018-2021 VMware, Inc.
+ * SPDX-License-Identifier: MIT
+ */
+
+export * from "./hierarchy"
